Refactor leaflet pinmaps into their own  components
